리뷰 조회, 리뷰 작성, 리뷰 평균 구하기 완료

+++ src/main/java/com/data/dataproject/dto/market/MarketDto.java
@@ -14,6 +14,6 @@ public class MarketDto {
     private String marketImage;
     private String marketName;
     private String marketAddress;
-    private Long review;
+    private Float review;
     private List<EventProduct> eventProducts;
 }
